+++ cauchy-services/src/lib.rs
@@ -1,2 +1,3 @@
+pub mod arena;
 pub mod database;
 pub mod peers;
